requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
ape-alchemy==0.8.11
ape-foundry==0.8.12
ape-vyper==0.8.10
asttokens==2.4.1
attrs==26.1.0
base58==1.0.3
bitarray==3.8.1
cached-property==2.0.1
cbor2==5.9.0
cchecksum==0.4.3
certifi==2026.2.25
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.1.8
cytoolz==1.1.0
decorator==5.2.1
eip712==0.3.3
eth-account==0.13.7
eth-ape==0.8.48
eth-bloom==3.1.0
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-pydantic-types==0.2.6
eth-rlp==2.2.0
eth-tester==0.13.0b1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
ethpm-types==0.6.30
evm-trace==0.2.6
evmchains==0.1.7
executing==2.2.1
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
ijson==3.5.0
immutables==0.21
iniconfig==2.3.0
ipython==8.39.0
jedi==0.19.2
lark==1.3.1
lazyasd==0.1.4
lru-dict==1.4.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
morphys==1.0
msgspec==0.21.0
multidict==6.7.1
numpy==1.26.4
packaging==23.2
pandas==2.3.3
parsimonious==0.10.0
parso==0.8.6
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cid==0.3.1
py-ecc==8.0.0
py-evm==0.12.1b1
py-geth==6.4.0
py-multibase==1.0.3
py-multicodec==0.2.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pymultihash==0.8.2
pytest==8.4.2
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==13.9.4
rlp==4.1.0
safe-pysha3==1.0.5
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.49
stack-data==0.6.3
toolz==1.1.0
tqdm==4.67.3
traitlets==5.14.3
trie==3.1.0
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
varint==1.0.2
vvm==0.3.2
vyper==0.4.3
watchdog==3.0.0
wcwidth==0.6.0
web3==7.15.0
websockets==15.0.1
wheel==0.45.1
yarl==1.23.0
